Move the example DFA porters into their own module
build_modulo_dfa / build_regex_dfa go from adapter.py to a new porters.py;
the package re-exports them, so consumers are unchanged.
